# frontend_rentelektro/src/types/profile.ts
import type { components } from '@/shared/api/generated/schema';

export type UserProfile = components['schemas']['UserProfile'];
export type UserProfileUpdate = components['schemas']['UserUpdate'];
export type PasswordChangePayload = components['schemas']['PasswordChangeRequest'];
export type AccountAnonymizePayload = components['schemas']['AccountAnonymizeRequest'];